chat/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
python-dotenv==1.0.1
openai==1.57.0
langchain==0.2.12
langchain-openai==0.1.20
langgraph==0.0.64

